Restructure blockdev module
